Доработка функционала вывода информации о дейтграмных пакетах

+++ tests/net/eth/socket.cpp
@@ -128,9 +128,9 @@ TEST_F(EthFixture, SocketMulticastTest){
 	ASSERT_NE(sock, awh::net::invalid_socket_t);
 
 	// Устанавливаем максимальное количество хопов для unicast пакетов в 1 (локальная сеть)
-	ASSERT_TRUE(this->_eth->socket.setHops(sock, awh::event::family_t::IPV4, awh::event::delivery_mode_t::UNICAST, awh::event::hops_t::NETWORK));
+	ASSERT_TRUE(this->_eth->socket.setHops(sock, awh::event::family_t::IPV4, awh::event::delivery_mode_t::UNICAST, static_cast <uint8_t> (awh::event::hops_t::NETWORK)));
 	// Получаем максимальное количество хопов для unicast пакетов
-	ASSERT_EQ(awh::event::hops_t::NETWORK, this->_eth->socket.getHops(sock, awh::event::family_t::IPV4, awh::event::delivery_mode_t::UNICAST));
+	ASSERT_EQ(static_cast <uint8_t> (awh::event::hops_t::NETWORK), this->_eth->socket.getHops(sock, awh::event::family_t::IPV4, awh::event::delivery_mode_t::UNICAST));
 
 	// Устанавливаем дифференцированные услуги (DSCP) для сокета
 	ASSERT_TRUE(this->_eth->socket.setDifferentiatedServicesCodePoint(sock, awh::event::family_t::IPV4, awh::event::dscp_t::CS0));
